Add a rudimentary Makefile for the DITA-OT toolchain
To save having to compile the documents manually, add a first stab
at a Makefile. Since the DITA-OT tools produce an overwhelming amount
of output on the console, redirect to a logfile per each file. Only
PDF is supported as of now simply to get started.

Also add a gitignore covering the buildlogs and the generated PDF.